requirements.txt
pandas 
sqlalchemy 
pymysql 
python-dotenv
cryptography
faker
fastapi 
uvicorn 
scikit-learn 
scipy